djangoApp/requirements.txt
django==4.0.2
pandas==1.4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
django-cors-headers==3.11.0
django-filter==21.1
django-bulk-update-or-create==0.3.0
ai-django-core==5.13.3
django-pandas==0.6.6
gunicorn==20.1.0
xlrd==2.0.1
openpyxl==3.0.9
psycopg2-binary==2.9.3
unicodecsv==0.14.1